feat: 예약 폼을 서버의 데이터와 일치 하도록 설정
