+++ src/apis/form.ts
@@ -1,0 +1,7 @@
+import { instance } from '@/apis/instance';
+import { UserSurveyRequest } from '@/types/api/formTypes';
+
+export const getFormList = async () => {
+  const res = await instance.get<UserSurveyRequest>('/form/list');
+  return res.data.data;
+};
